review: move old test to new impl + additional test
